Split out waiter and collector in separate files
